iox-#605 Move RelativePointer from internal to public include directory
